[rtl] Replace peripheral's direct access to XXX_base_c constants with a generic
This won't change generated netlist and doesn't truly allow to select a different
base address that is already defined as a constant in neorv32_package because
address decoding is made using the whole 32 bits words, so changing just the base
address won't make any module recognize access to its own registers.

Next step should be processing only the lower bits of the address to decide
on which register to act upon. Once this is done, it should be much easier
adding different devices of the same kind to the SoC. There are already two
UARTs that would be best candidates to test this concept on.