[rtl] SPI: replace hardcoded register addresses with an offset declared inside the module
Since the peripheral is already checking the full address to see if it
matches its address space, there's no need for it to check it again when
comparing incoming address with register address. Instead, only the lower
bits can be checked and the registers full address replaced with just the
offset inside the address space that belongs to the peripheral.

Also, addresses (or offsets) of registers are internal of the module and
they do not need to be declared inside the global package, so remove them
from neorv32_package.vhd

This also seems to reduce the quantity of logic used (as expected, given
that less bits are used in the comparisons explained above).

--- build/yosys-report.txt      2023-04-17 11:04:44.676147942 +0200
+++ build.new/yosys-report.txt  2023-04-17 10:55:54.804638843 +0200
 === art_gs_top ===

-   Number of wires:               3464
-   Number of wire bits:          30416
-   Number of public wires:        3464
-   Number of public wire bits:   30416
+   Number of wires:               3467
+   Number of wire bits:          30360
+   Number of public wires:        3467
+   Number of public wire bits:   30360
    Number of memories:               0
    Number of memory bits:            0
    Number of processes:              0
-   Number of cells:               5437
+   Number of cells:               5427
      $_TBUF_                         2
      SB_CARRY                      488
      SB_DFF                        232
@@ -3304,7 +3302,7 @@
      SB_DFFS                         2
      SB_DFFSR                      262
      SB_HFOSC                        1
-     SB_LUT4                      3159
+     SB_LUT4                      3149
      SB_PLL40_CORE                   1
      SB_RAM40_4K                    12
      SB_SPRAM256KA                   4